refactor(app): extract abort + pending registry from submit.ts
Move the AbortSource/PendingPrompt types, the shared pending map, and the
abort closure out of prompt-input/submit.ts into a new submit-abort.ts as a
createAbort(deps) factory. Pure extraction, no behavior change; submit.ts
imports the pending registry and wires createAbort with its existing deps.
